# api.yaml
openapi: 3.0.0
servers:
    - url: http://localhost:3000/api
info:
    title: Bulldawg Marketplace API
    description: API for Bulldawg Marketplace Backend Functions
    version: 1.0.0
paths:
    /ping:
        get:
            summary: Server Ping
            description: Ensures that the API is functioning properly
            responses:
                '200':
                    description: Pong
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: integer
                                        description: Whether the request was successful or not
                                        example: 1
                                    message:
                                        type: string
                                        description: A message indicating the server response
                                        example: 'Pong!'
    /login:
        post:
            summary: Login
            description: Logs in using email/username and password.
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    description: 'email@example.com'
                                password:
                                    type: string
                                    description: 'password123'
            responses:
                '400':
                    description: Invalid JSON body
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            examples:
                                MissingInfo:
                                    value:
                                        success: 0
                                        message: 'Either email or username is required.'
                                        error: 1
                                InvalidJSON:
                                    value:
                                        success: 0
                                        message: 'Error parsing JSON.'
                                        error: 1
                '401':
                    description: Invalid credentials
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            example:
                                success: 0
                                message: 'The password is invalid.'
                                error: 1
                '200':
                    description: Successful login
                    headers:
                        Set-Cookie:
                            description: Authorization token to be used for future requests
                            schema:
                                type: string
                                example: token=<token>; Path=/; HttpOnly; SameSite=Strict; MaxAge=3600;
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            example:
                                success: 1
                                message: 'Login Successful'
                                error: 0
    /register:
        post:
            summary: Register
            description: Registers a new account.
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    description: 'email@example.com'
                                username:
                                    type: string
                                    description: 'exampleUser'
                                password:
                                    type: string
                                    description: 'password123'
                                firstName:
                                    type: string
                                    description: 'John'
                                lastName:
                                    type: string
                                    description: 'Doe'
                                profilePic:
                                    type: string
                                    description: 'Profile Photo URL (Optional)'
            responses:
                '400':
                    description: Invalid JSON body
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            examples:
                                MissingInfo:
                                    value:
                                        success: 0
                                        message: 'email, username, password, firstName, & lastName fields are required!'
                                        error: 1
                                InvalidJSON:
                                    value:
                                        success: 0
                                        message: 'Error parsing JSON.'
                                        error: 1
                                ExistingUser:
                                    value:
                                        success: 0
                                        message: '(Username <username>)/(Email <email>) is already in use.'
                                        error: 1
                '200':
                    description: Successful registration
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            examples:
                                Success:
                                    value:
                                        success: 1
                                        message: 'Registered successfully!'
                                        error: 0
    /post/create:
        post:
            summary: Create a Listing
            description: Allows a user to list a new product for sale.
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                    example: Mercedes Benz Car
                                price:
                                    type: number
                                    example: 43000
                                description:
                                    type: string
                                    example: A lightly used German racing machine
                                pic:
                                    type: string
                                    example: A URL to a photo for the listing
            security:
                - cookieAuth: []
            responses:
                '201':
                    description: 'Successful Listing'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            examples:
                                Success:
                                    value:
                                        success: 1
                                        message: 'Post created successfully!'
                                        error: 0
                '401':
                    description: 'Unauthorized Request'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            examples:
                                Unauthorized:
                                    value:
                                        success: 0
                                        message: 'Unauthorized Request.'
                                        error: 0

                '400':
                    description: Invalid JSON body
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Response'
                            examples:
                                MissingInfo:
                                    value:
                                        success: 0
                                        message: 'title, price, pic, & description fields are required.'
                                        error: 1
                                InvalidJSON:
                                    value:
                                        success: 0
                                        message: 'Error parsing JSON.'
                                        error: 1
    /me:
        get:
            summary: Me
            description: Checks authentication status for given user.
            security:
                - cookieAuth: []
            responses:
                '200':
                    description: 'Successfully Authenticated'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    email:
                                        type: string
                                        example: 'email@example.com'
                                    username:
                                        type: string
                                        example: 'exampleUser'
                '401':
                    description: 'Unauthorized'
                    content:
                        application/json:
                            schema:
                                type: object
    /logout:
        get:
            summary: Logout
            description: Destroys session cookie and logs the current user out.
            security:
                - cookieAuth: []
            responses:
                '200':
                    description: 'Logout'
                    headers:
                        Set-Cookie:
                            description: Destroyed authorization cookie
                            schema:
                                type: string
                                example: token=; Path=/; HttpOnly; SameSite=Strict; MaxAge=-1;

components:
    schemas:
        Response:
            type: object
            properties:
                success:
                    description: Whether request was successful.
                    type: integer
                    enum:
                        - 0
                        - 1
                message:
                    description: Response message
                    type: string
                error:
                    description: Whether an error occurred.
                    type: integer
                    enum:
                        - 0
                        - 1
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: token